MultiThreading

+++ EntityFW_ConsoleApp/MultiThreading/Classes/RandomIntArrayGenerator.cs
@@ -8,15 +8,15 @@ namespace MultiThreading.Classes
 {
     public static class RandomIntArrayGenerator
     {
-        public static uint[] Generate(int size, int minValue, int maxValue)
+        public static long[] Generate(int size, int minValue, int maxValue)
         {
-            uint[] result = new uint[size];
+            long[] result = new long[size];
 
             Random rnd = new Random();
 
             for (int i = 0; i < size; i++)
             {
-                result[i] = Convert.ToUInt32(rnd.Next(minValue, maxValue + 1));
+                result[i] = Convert.ToInt64(rnd.Next(minValue, maxValue + 1));
             }
 
             return result;
